Fix datatype, validation

+++ app/src/main/java/com/yellion/yellapp/BudgetStatisticIncomeFragment.java
@@ -81,7 +81,7 @@ public class BudgetStatisticIncomeFragment extends Fragment {
         binding.idCreateDate.setText(createDate);
         binding.threshold.setText(String.valueOf(budgetCard.getThreshold()));
         binding.balance2.setText(String.valueOf(budgetCard.getBalance()));
-        int progress = budgetCard.getBalance()/budgetCard.getThreshold()*100;
+        int progress = (int)(budgetCard.getBalance()/budgetCard.getThreshold()*100);
         if (progress > 100) progress = 100;
         binding.circularProgressbar.setProgress(progress);
         binding.tv.setText(String.valueOf(progress)+"%");
